Created ProcessManager class.
It exposes a Process interface that is responsible of holding
process information.

The ProcessManager class needs a Process object and handles
creation of detached processes, saving or deletion of their information,
kill and loading.

The loadProcess static function returns an instance of ProcessManager
with loaded -from file- information about the process itself.

The class needs to be tested.