Added handler for server status cli command.
The process will work with a post entrypoint that requires a json body
with a token variable, representing the security.token string in
config.yml.
The post request, if successfull, will return the pid of the process
of the server.

The status command will thus send a request to the /ping entrypoint:
if a response is received, the server is online;
it is offline otherwise.