Rebuilt frontend base boilerplate, using typescript.
Better file structure has been used.
Built the PageLoadingBar component, that can be found under
components/page-loading-bar.
It will be triggered automatically when lazy-calling routes.

Hmr has been added, but it suffers from a bug that will
reload the entire page instead of replacing the module.
The issue is not exactly identified, and it can be found in lazy route
loading, or typescript transpile.
The hmr setup will run only in debug mode.

The log class has been extended with the RFC5424 syslog levels.
The debug level can be seen only in debug mode, using --debug when
starting the server.

The circleCI environment detection has been changed using the CIRCLECI
environment variable provided by the service itself.

The test suite is gaunt: it will be implemented as soon as possible.